Menu Transitions and GameMap
Menus are temporarily switched to not use LoadingScreen because of
incompletion of transition on some machines.

\Screens\InnerGame has been created.  In it should go screens
responsible for displaying all content not related to the Main Menu.

The ScenarioScreen class is responsible for displaying the
action-oriented Scenario phase of the game.

We have renamed Map to GameMap to prevent its confusion with the Map
class.